# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Owl Invites API",
    "version": "0.1.0",
    "description": "Gate 2 identity, event membership, invitation household, and RSVP capability contract."
  },
  "servers": [{ "url": "/api/v1" }],
  "components": {
    "securitySchemes": {
      "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "session" },
      "invitationSessionCookie": { "type": "apiKey", "in": "cookie", "name": "owl_invitation_session" }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "required": ["id", "email", "name", "timezone", "instanceRole", "status", "isAdmin", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "name": { "type": "string" },
          "timezone": { "type": "string" },
          "instanceRole": { "type": "string", "enum": ["admin", "user"] },
          "status": { "type": "string", "enum": ["invited", "active", "disabled"] },
          "isAdmin": { "type": "boolean", "deprecated": true },
          "invitedByUserId": { "type": "string", "format": "uuid" },
          "activatedAt": { "type": "string", "format": "date-time" },
          "lastLoginAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "InstanceSettings": {
        "type": "object",
        "required": ["instanceName", "defaultTimezone", "allowSignups", "supportEmail", "configured"],
        "properties": {
          "instanceName": { "type": "string", "maxLength": 200 },
          "defaultTimezone": { "type": "string", "maxLength": 100 },
          "allowSignups": { "type": "boolean" },
          "supportEmail": { "type": "string" },
          "configured": { "type": "boolean", "readOnly": true }
        }
      },
      "InstanceSettingsUpdate": {
        "type": "object",
        "additionalProperties": false,
        "required": ["instanceName", "defaultTimezone", "allowSignups", "supportEmail"],
        "properties": {
          "instanceName": { "type": "string", "maxLength": 200 },
          "defaultTimezone": { "type": "string", "maxLength": 100 },
          "allowSignups": { "type": "boolean" },
          "supportEmail": { "type": "string" }
        }
      },
...
      "AuditEntry": {
        "type": "object",
        "required": ["id", "actorKind", "action", "metadata", "createdAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "actorUserId": { "type": "string", "format": "uuid" },
          "actorKind": { "type": "string", "enum": ["user", "cli", "system"] },
          "action": { "type": "string" },
          "targetUserId": { "type": "string", "format": "uuid" },
          "eventId": { "type": "string", "format": "uuid" },
          "metadata": { "type": "object", "additionalProperties": true },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "Event": {
        "type": "object",
        "required": ["id", "organizerId", "title", "description", "eventDate", "location", "timezone", "retentionDays", "status", "shareToken", "contactRequirement", "showHeadcount", "showGuestList", "waitlistEnabled", "commentsEnabled", "seriesOverride", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "organizerId": { "type": "string", "format": "uuid", "deprecated": true },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "eventDate": { "type": "string", "format": "date-time" },
          "endDate": { "type": "string", "format": "date-time" },
          "location": { "type": "string" },
          "timezone": { "type": "string" },
          "retentionDays": { "type": "integer" },
          "status": { "type": "string", "enum": ["draft", "published", "cancelled", "archived"] },
          "shareToken": { "type": "string" },
          "contactRequirement": { "type": "string", "enum": ["email", "phone", "email_or_phone", "email_and_phone"] },
          "showHeadcount": { "type": "boolean" },
          "showGuestList": { "type": "boolean" },
          "rsvpDeadline": { "type": "string", "format": "date-time" },
          "maxCapacity": { "type": "integer" },
          "waitlistEnabled": { "type": "boolean" },
          "commentsEnabled": { "type": "boolean" },
          "seriesId": { "type": "string", "format": "uuid" },
          "seriesIndex": { "type": "integer" },
          "seriesOverride": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Invitation": {
        "type": "object",
        "required": ["id", "eventId", "label", "preferredDeliveryMethod", "additionalGuestAllowance", "source", "tokenVersion", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "eventId": { "type": "string", "format": "uuid" },
          "label": { "type": "string", "maxLength": 200 },
          "contactEmail": { "type": "string", "format": "email" },
          "contactPhone": { "type": "string" },
          "preferredDeliveryMethod": { "type": "string", "enum": ["email", "sms", "none"] },
          "additionalGuestAllowance": { "type": "integer", "minimum": 0 },
          "source": { "type": "string", "enum": ["private", "open"] },
          "tokenVersion": { "type": "integer", "minimum": 1 },
          "revokedAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "InvitationGuest": {
        "type": "object",
        "required": ["id", "invitationId", "name", "origin", "attendance"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "invitationId": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "origin": { "type": "string", "enum": ["assigned", "additional"] },
          "attendance": { "type": "string", "enum": ["pending", "attending", "maybe", "declined"] }
        }
      },
      "InvitationQuestion": {
        "type": "object",
        "required": ["id", "label", "type", "options", "required", "scope", "sortOrder"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "label": { "type": "string" },
          "type": { "type": "string", "enum": ["text", "select", "checkbox"] },
          "options": { "type": "array", "items": { "type": "string" } },
          "required": { "type": "boolean" },
          "scope": { "type": "string", "enum": ["invitation", "guest"] },
          "sortOrder": { "type": "integer" }
        }
      },
      "InvitationHousehold": {
        "type": "object",
        "required": ["invitation", "event", "response", "guests", "questions", "invitationAnswers", "guestAnswers"],
        "properties": {
          "invitation": { "$ref": "#/components/schemas/Invitation" },
          "event": { "type": "object", "additionalProperties": true },
          "response": { "type": "object", "additionalProperties": true },
          "guests": { "type": "array", "items": { "$ref": "#/components/schemas/InvitationGuest" } },
          "questions": { "type": "array", "items": { "$ref": "#/components/schemas/InvitationQuestion" } },
          "invitationAnswers": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "guestAnswers": { "type": "array", "items": { "type": "object", "additionalProperties": true } }
        }
      },
      "CreateInvitationRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["label", "preferredDeliveryMethod", "additionalGuestAllowance", "assignedGuestNames"],
        "properties": {
          "label": { "type": "string", "maxLength": 200 },
          "contactEmail": { "type": "string", "format": "email" },
          "contactPhone": { "type": "string" },
          "preferredDeliveryMethod": { "type": "string", "enum": ["email", "sms", "none"] },
          "additionalGuestAllowance": { "type": "integer", "minimum": 0, "maximum": 50 },
          "assignedGuestNames": { "type": "array", "minItems": 1, "items": { "type": "string" } },
          "send": { "type": "boolean" }
        }
      },
...
    }
  },
  "paths": {
...
    "/setup/config": {
      "get": {
        "operationId": "getInstanceSettings",
        "security": [{ "sessionCookie": [] }],
        "responses": { "200": { "description": "Settings", "content": { "application/json": { "schema": { "type": "object", "required": ["data"], "properties": { "data": { "$ref": "#/components/schemas/InstanceSettings" } } } } } } }
      },
      "post": {
        "operationId": "updateInstanceSettings",
        "security": [{ "sessionCookie": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InstanceSettingsUpdate" } } } },
        "responses": { "200": { "description": "Updated settings", "content": { "application/json": { "schema": { "type": "object", "required": ["data"], "properties": { "data": { "$ref": "#/components/schemas/InstanceSettings" } } } } } } }
      }
    },
...
    "/admin/users": {
      "get": {
        "operationId": "listUsers",
        "security": [{ "sessionCookie": [] }],
        "responses": { "200": { "description": "Users", "content": { "application/json": { "schema": { "type": "object", "required": ["data"], "properties": { "data": { "type": "object", "required": ["users"], "properties": { "users": { "type": "array", "items": { "$ref": "#/components/schemas/User" } } } } } } } } } }
      }
    },
...
    "/admin/audit": {
      "get": {
        "operationId": "listAdminAudit",
        "security": [{ "sessionCookie": [] }],
        "parameters": [{ "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 200 } }],
        "responses": { "200": { "description": "Audit entries", "content": { "application/json": { "schema": { "type": "object", "required": ["data"], "properties": { "data": { "type": "object", "required": ["entries"], "properties": { "entries": { "type": "array", "items": { "$ref": "#/components/schemas/AuditEntry" } } } } } } } } } }
      }
    },
...
    "/events": {
      "get": {
        "operationId": "listEvents",
        "security": [{ "sessionCookie": [] }],
        "responses": { "200": { "description": "Explicitly accessible events", "content": { "application/json": { "schema": { "type": "object", "required": ["data"], "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Event" } } } } } } } }
      }
    },
...
    "/events/{eventId}/invitations": {
      "parameters": [{ "name": "eventId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
      "get": {
        "operationId": "listInvitations",
        "security": [{ "sessionCookie": [] }],
        "responses": { "200": { "description": "Invitation households for the event", "content": { "application/json": { "schema": { "type": "object", "required": ["data"], "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/InvitationHousehold" } } } } } } } }
      },
      "post": {
        "operationId": "createInvitation",
        "security": [{ "sessionCookie": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateInvitationRequest" } } } },
        "responses": { "201": { "description": "Private invitation created" } }
      }
    },
...
  }
}