# internal/server/router.go
package server

import (
	"context"
	"encoding/json"
	"io/fs"
	"net/http"
	"path/filepath"
	"strings"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/go-chi/cors"
	"github.com/rs/zerolog"

	apidoc "github.com/yannkr/openrsvp/api"
	"github.com/yannkr/openrsvp/internal/security"
)

// routes builds and returns the chi router with all middleware and routes.
func (s *Server) routes() *chi.Mux {
	r := chi.NewRouter()

	// --- Middleware ---
	// Resolve client identity and forwarded scheme before any middleware uses
	// them. Forwarded headers from untrusted peers are stripped fail-closed.
	r.Use(security.TrustedProxyMiddleware(s.cfg.TrustedProxies))
	// Set baseline security headers on every response.
	r.Use(security.SecurityHeadersMiddleware())
	r.Use(cors.Handler(cors.Options{
		AllowedOrigins:   []string{s.cfg.BaseURL},
		AllowedMethods:   []string{"GET", "POST", "PUT", "PATCH", "DELETE", "OPTIONS"},
		AllowedHeaders:   []string{"Accept", "Authorization", "Content-Type", "X-CSRF-Token"},
		ExposedHeaders:   []string{"Link"},
		AllowCredentials: true,
		MaxAge:           300,
	}))
	r.Use(middleware.Compress(5, "text/html", "text/css", "application/javascript", "application/json", "image/svg+xml", "text/plain"))
	r.Use(zerologMiddleware(s.logger))
	r.Use(middleware.Recoverer)
	r.Use(middleware.RequestID)
	r.Use(s.securityMw.CSRF)

	// --- Health checks ---
	r.Get("/health", s.handleHealth)
	r.Get("/health/ready", s.handleHealthReady)

	// --- API v1 ---
	r.Route("/api/v1", func(api chi.Router) {
		// General rate limiting applies to API routes only (not static SPA files).
		api.Use(security.RateLimitMiddleware(s.securityMw.GeneralRateLimiter))
		// Limit request body size to 1 MB for API routes.
		api.Use(security.BodyLimitMiddleware(1 << 20))
		// Sanitize all incoming JSON request bodies.
		api.Use(s.securityMw.Sanitize)

		api.Get("/health", s.handleHealth)
		api.Get("/openapi.json", apidoc.ServeHTTP)

		// Public app config (non-sensitive feature flags).
		api.Get("/config", func(w http.ResponseWriter, r *http.Request) {
			w.Header().Set("Content-Type", "application/json")
			_ = json.NewEncoder(w).Encode(map[string]any{
				"data": map[string]any{
					"smsEnabled": s.cfg.SMSEnabled(),
				},
			})
		})

		// Auth routes: the stricter rate limiter (10/min) is passed into
		// Routes() and applied only to magic-link and verify.  Session
		// management endpoints (/me, /logout) fall under the general API
		// rate limiter so SPA page-load calls to /auth/me don't exhaust
		// the auth rate budget.
		api.Mount("/auth/account-invites", s.userAdminHandler.PublicRoutes(
			security.RateLimitMiddleware(s.securityMw.AuthRateLimiter),
		))
		api.Mount("/auth", s.authHandler.Routes(
			security.RateLimitMiddleware(s.securityMw.AuthRateLimiter),
		))

		// Series routes must be mounted before event routes so that
		// /events/series is not captured by the event handler's /{eventId} pattern.
		api.Mount("/events/series", s.seriesHandler.Routes())
		api.Route("/events/{eventId}", func(eventRoutes chi.Router) {
			eventRoutes.Mount("/", s.invitationHandler.OrganizerRoutes())
		})
		api.Mount("/events", s.eventHandler.Routes())
		api.Mount("/invitations", s.invitationHandler.PublicRoutes())

		// Question routes nested under events (organizer-only).
		api.Route("/events/{eventId}/questions", func(qr chi.Router) {
			qr.Mount("/", s.questionHandler.Routes())
		})

		// RSVP routes with moderate rate limiting (30/min) and honeypot on public submissions.
		api.Route("/rsvp", func(rsvpR chi.Router) {
			rsvpR.Use(security.RateLimitMiddleware(s.securityMw.RSVPRateLimiter))
			rsvpR.Use(s.securityMw.Honeypot)
			rsvpR.Mount("/", s.rsvpHandler.Routes())
		})

		api.Mount("/invite", s.inviteHandler.Routes())

		// Serve uploaded files (public, for shared invite pages).
		uploadsPrefix := "/uploads/"
		api.Get(uploadsPrefix+"*", func(w http.ResponseWriter, r *http.Request) {
			// Strip prefix to get filename, then take only the base name
			// to prevent path traversal attacks (e.g. ../../etc/passwd).
			name := filepath.Base(strings.TrimPrefix(r.URL.Path, "/api/v1"+uploadsPrefix))

			// Security headers to prevent MIME-sniffing polyglot file attacks.
			// nosniff stops browsers from guessing a different Content-Type.
			w.Header().Set("X-Content-Type-Options", "nosniff")
			// Strict CSP blocks any script execution even if the browser
			// renders the file as HTML.
			w.Header().Set("Content-Security-Policy", "default-src 'none'")
			// Prevent the uploaded file from being embedded in a frame.
			w.Header().Set("X-Frame-Options", "DENY")

			http.ServeFile(w, r, filepath.Join(s.uploadsDir, name))
		})
		api.Mount("/messages", s.messageHandler.Routes())
		api.Mount("/reminders", s.reminderHandler.Routes())
		api.Mount("/feedback", s.feedbackHandler.Routes())
		api.Mount("/comments", s.commentHandler.Routes())
		api.Mount("/webhooks", s.webhookHandler.Routes())
		api.Mount("/notifications", s.notifHandler.Routes())
		api.Mount("/admin/users", s.userAdminHandler.UserRoutes())
		api.Mount("/admin/audit", s.userAdminHandler.AuditRoutes())
		api.Mount("/admin/events", s.eventAdminHandler.Routes())
		api.Mount("/admin", s.statsHandler.Routes())
		// Public, token-based email unsubscribe (no auth, CSRF-exempt).
		api.Mount("/unsubscribe", s.suppressionHandler.Routes())
		// Instance setup: status and one-time token bootstrap are public;
		// ongoing config remains admin-only.
		api.Mount("/setup", s.instanceConfigHandler.Routes(
			security.RateLimitMiddleware(s.securityMw.AuthRateLimiter),
		))
	})

	// --- Static files / SPA fallback ---
	s.mountStaticFiles(r)

	return r
}

// handleHealth returns a simple 200 OK with status information.
func (s *Server) handleHealth(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	_ = json.NewEncoder(w).Encode(map[string]string{
		"status": "ok",
	})
}

// handleHealthReady returns 200 if the database is reachable, 503 otherwise.
func (s *Server) handleHealthReady(w http.ResponseWriter, r *http.Request) {
	ctx, cancel := context.WithTimeout(r.Context(), 3*time.Second)
	defer cancel()

	var result int
	err := s.db.QueryRowContext(ctx, "SELECT 1").Scan(&result)

	w.Header().Set("Content-Type", "application/json")

	if err != nil {
		s.logger.Error().Err(err).Msg("health check: database unreachable")
		w.WriteHeader(http.StatusServiceUnavailable)
		_ = json.NewEncoder(w).Encode(map[string]string{
			"status":   "unavailable",
			"database": "unreachable",
		})
		return
	}

	w.WriteHeader(http.StatusOK)
	_ = json.NewEncoder(w).Encode(map[string]string{
		"status":   "ok",
		"database": "connected",
	})
}

// mountStaticFiles serves embedded frontend assets with SPA fallback.
func (s *Server) mountStaticFiles(r *chi.Mux) {
	staticFS := getFrontendFS()

	if staticFS != nil {
		fileServer := http.FileServer(http.FS(staticFS))

		// Pre-read the SPA fallback page (200.html) for client-side routing.
		// This is separate from index.html so that prerendered pages
		// (like the landing page) aren't overwritten by the SPA shell.
		fallbackHTML, _ := fs.ReadFile(staticFS, "200.html")

		r.NotFound(func(w http.ResponseWriter, r *http.Request) {
			// Try to serve the actual file first.
			path := r.URL.Path[1:] // strip leading /
			if path == "" {
				path = "index.html"
			}
			f, err := staticFS.Open(path)
			if err == nil {
				info, statErr := f.Stat()
				_ = f.Close()
				if statErr == nil && !info.IsDir() {
					// Vite-hashed assets are safe to cache forever.
					if strings.HasPrefix(r.URL.Path, "/_app/immutable/") {
						w.Header().Set("Cache-Control", "public, max-age=31536000, immutable")
					} else {
						w.Header().Set("Cache-Control", "no-cache")
					}
					fileServer.ServeHTTP(w, r)
					return
				}
			}
			// SPA fallback: serve index.html directly for client-side routing.
			w.Header().Set("Content-Type", "text/html; charset=utf-8")
			w.Header().Set("Cache-Control", "no-cache")
			w.WriteHeader(http.StatusOK)
			_, _ = w.Write(fallbackHTML)
		})
	} else {
		r.NotFound(func(w http.ResponseWriter, r *http.Request) {
			w.Header().Set("Content-Type", "application/json")
			w.WriteHeader(http.StatusNotFound)
			_ = json.NewEncoder(w).Encode(map[string]string{
				"error": "not found",
			})
		})
	}
}

// zerologMiddleware returns a chi middleware that logs requests using zerolog.
func zerologMiddleware(logger zerolog.Logger) func(http.Handler) http.Handler {
	return func(next http.Handler) http.Handler {
		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
			start := time.Now()
			ww := middleware.NewWrapResponseWriter(w, r.ProtoMajor)

			defer func() {
				logger.Info().
					Str("method", r.Method).
					Str("path", r.URL.Path).
					Int("status", ww.Status()).
					Int("bytes", ww.BytesWritten()).
					Dur("duration", time.Since(start)).
					Str("remote", r.RemoteAddr).
					Str("request_id", middleware.GetReqID(r.Context())).
					Msg("request")
			}()

			next.ServeHTTP(ww, r)
		})
	}
}